Updated the Application file and added the MovieDBBuilder file again. Separated out the RecyclerViewAdapter to two.
